feat(cli): help improvements
+ refactoring:
  cli validation moved from app-helper.js to base-cli-handler.js
+ help for one and help for all parameters functions

Closes ENG-556
