+++ src/cli/base-cli-handler.js
@@ -13,6 +13,9 @@
 
 const commandLineArgs = require('command-line-args');
 const commandLineUsage = require('command-line-usage');
+const AppHelper = require('../helpers/app-helper');
+const Errors = require('../helpers/errors');
+const ErrorMessages = require('../helpers/error-messages');
 
 class CLIHandler {
   constructor() {
@@ -29,9 +32,44 @@ class CLIHandler {
     return commandLineArgs(commandDefinitions, Object.assign({camelCase: true, partial: true,}, options))
   }
 
-  help(hide = [], showOptions = true, hasCommands = true, additionalSection = []) {
+  help(show = [], showOptions = true, hasCommands = true, additionalSection = []) {
+
+    if (show.length === 0) {
+      //show all
+      this.helpAll(show, showOptions, hasCommands, additionalSection)
+    } else {
+      //show list
+      this.helpSome(show, showOptions)
+    }
+  }
+
+  helpSome(show = [], showOptions = true,) {
     const options = Object.keys(this.commands)
-      .filter((key) => hide.indexOf(key) === -1)
+      .filter((key) => show.indexOf(key) !== -1)
+      .map((key) => ({
+        header: key,
+        optionList: this.commandDefinitions,
+        group: [key],
+      }));
+
+    const sections = [
+      {
+        header: 'Usage',
+        content: `$ iofog-controller ${this.name} ${show.length === 1 ? show : '<command>'} <options>`,
+      },
+    ].concat(showOptions ? options : []);
+
+    const usage = [
+      {
+        header: 'ioFogController',
+        content: 'Fog Controller project for Eclipse IoFog @ iofog.org \\nCopyright (c) 2018 Edgeworx, Inc.',
+      }
+    ].concat(sections);
+    console.log(commandLineUsage(usage))
+  }
+
+  helpAll(show = [], showOptions = true, hasCommands = true, additionalSection = []) {
+    const options = Object.keys(this.commands)
       .map((key) => ({
         header: key,
         optionList: this.commandDefinitions,
@@ -62,6 +100,173 @@ class CLIHandler {
     ].concat(sections);
     console.log(commandLineUsage(usage))
   }
+
+  handleCLIError(error) {
+    switch (error.name) {
+      case "UNKNOWN_OPTION":
+        console.log("Invalid argument '" + error.optionName.split('-').join('') + "'");
+        break;
+      case "UNKNOWN_VALUE":
+        console.log("Invalid value " + error.value);
+        break;
+      case "InvalidArgumentError":
+        console.log(error.message);
+        break;
+      case "InvalidArgumentTypeError":
+        console.log(error.message);
+        break;
+      case "ALREADY_SET":
+        console.log("Parameter '" + error.optionName + "' is used multiple times");
+        break;
+      default:
+        console.log(JSON.stringify(error));
+        break;
+    }
+  }
+
+  validateParameters(command, commandDefinitions, args) {
+    // 1st argument = command
+    args.shift();
+
+    const possibleAliasesList = _getPossibleAliasesList(command, commandDefinitions);
+    const possibleArgsList = _getPossibleArgsList(command, commandDefinitions);
+
+    let expectedValueType;
+    let currentArgName;
+
+    if (args.length === 0) {
+      return
+    }
+    const argsMap = argsArrayAsMap(args);
+
+    argsMap.forEach((values, key) => {
+      if (key.startsWith("--")) { // argument
+        // '--ssl-cert' format -> 'ssl-cert' format
+        const argument = key.substr(2);
+        _validateArg(argument, possibleArgsList);
+        currentArgName = argument;
+        expectedValueType = _getValType(argument, commandDefinitions);
+      } else if (key.startsWith("-")) { // alias
+        // '-q' format -> 'q' format
+        const alias = key.substr(1);
+        _validateArg(alias, possibleAliasesList);
+        currentArgName = alias;
+        expectedValueType = _getValType(alias, commandDefinitions);
+      }
+
+      let valType;
+      if (values.length === 0) {
+        valType = 'boolean';
+      } else if (values.length === 1) {
+        const firstVal = Number(values[0]);
+        if (Number.isNaN(firstVal.valueOf())) {
+          valType = 'string';
+        } else if (Number.isInteger(firstVal.valueOf())) {
+          valType = 'integer';
+        } else {
+          valType = 'float'
+        }
+      }
+      //TODO else validate multiply parameters. Add after multiply parameters will be used in cli api
+
+      let isValidType = true;
+      if (expectedValueType === 'string' && valType === 'boolean') {
+        isValidType = false;
+      } else if ((expectedValueType === 'float' || expectedValueType === 'number')
+        && (valType !== 'float' && valType !== 'number' && valType !== 'integer')) {
+        isValidType = false;
+      } else if (expectedValueType === 'integer' && valType !== 'integer') {
+        isValidType = false;
+      } else if (expectedValueType === 'boolean' && valType !== 'boolean') {
+        isValidType = false;
+      }
+
+      if (!isValidType) {
+        throw new Errors.InvalidArgumentTypeError(AppHelper.formatMessage(ErrorMessages.INVALID_CLI_ARGUMENT_TYPE, currentArgName, expectedValueType));
+      }
+    })
+  }
+}
+
+function argsArrayAsMap(args) {
+  let argsVars = args.join(' ').split(/(?= -{1,2}[^-]+)/);
+  const argsMap = new Map();
+  argsVars
+    .map(pair => pair.trim())
+    .map(pair => {
+      const spaceIndex = pair.indexOf(' ');
+      let key, values;
+      if (spaceIndex !== -1) {
+        key = pair.substr(0, pair.indexOf(' '));
+        values = pair.substr(pair.indexOf(' ')+1).split(' ');
+        argsMap.set(key, values);
+      } else {
+        key = pair;
+        values = [];
+      }
+      argsMap.set(key, values);
+
+    });
+  return argsMap;
+}
+
+function _validateArg(arg, aliasesList) {
+  const valid = aliasesList.includes(arg);
+  if (!valid) {
+    throw new Errors.InvalidArgumentError("Invalid argument '" + arg + "'");
+  }
+}
+
+function _getPossibleAliasesList(command, commandDefinitions) {
+  const possibleAliasesList = [];
+
+  for (const definition of commandDefinitions) {
+    const group = definition.group;
+    const isGroupArray = group.constructor === Array;
+    if (isGroupArray) {
+      for (const gr of group) {
+        if (gr === command) {
+          possibleAliasesList.push(definition.alias);
+          break;
+        }
+      }
+    } else {
+      if (group === command) {
+        possibleAliasesList.push(definition.alias);
+      }
+    }
+  }
+
+  return possibleAliasesList;
+}
+
+function _getPossibleArgsList(command, commandDefinitions) {
+  const possibleArgsList = [];
+
+  for (const definition of commandDefinitions) {
+    const group = definition.group;
+    const isGroupArray = group.constructor === Array;
+    if (isGroupArray) {
+      for (const gr of group) {
+        if (gr === command) {
+          possibleArgsList.push(definition.name);
+          break;
+        }
+      }
+    } else {
+      if (group === command) {
+        possibleArgsList.push(definition.name);
+      }
+    }
+  }
+
+  return possibleArgsList;
+}
+
+function _getValType(arg, commandDefinitions) {
+  const command = commandDefinitions
+    .filter(def => def.name === arg || def.alias === arg)[0];
+  return command.type.name.toLowerCase();
 }
 
 module.exports = CLIHandler;
