+++ src/cli/catalog.js
@@ -158,7 +158,7 @@ class Catalog extends BaseCLIHandler {
 
       const command = catalogCommand.command.command;
 
-      AppHelper.validateParameters(command, this.commandDefinitions, args.argv);
+      this.validateParameters(command, this.commandDefinitions, args.argv);
 
       switch (command) {
         case constants.CMD_ADD:
@@ -181,12 +181,12 @@ class Catalog extends BaseCLIHandler {
           return this.help()
       }
     } catch (error) {
-      AppHelper.handleCLIError(error);
+      this.handleCLIError(error);
     }
   }
 
   help() {
-    super.help([constants.CMD_LIST], true, true, [
+    super.help([], true, true, [
       {
         header: 'JSON File Schema',
         content: [
