+++ src/cli/flow.js
@@ -87,7 +87,7 @@ class Flow extends BaseCLIHandler {
 
       const command = flowCommand.command.command;
 
-      AppHelper.validateParameters(command, this.commandDefinitions, args.argv);
+      this.validateParameters(command, this.commandDefinitions, args.argv);
 
       switch (command) {
         case constants.CMD_ADD:
@@ -110,12 +110,12 @@ class Flow extends BaseCLIHandler {
           return this.help([constants.CMD_LIST])
       }
     } catch (error) {
-      AppHelper.handleCLIError(error);
+      this.handleCLIError(error);
     }
   }
 
   help() {
-    super.help([constants.CMD_LIST], true, true, [
+    super.help([], true, true, [
       {
         header: 'JSON File Schema',
         content: [
