+++ src/cli/registry.js
@@ -94,7 +94,7 @@ class Registry extends BaseCLIHandler {
 
       const command = registryCommand.command.command;
 
-      AppHelper.validateParameters(command, this.commandDefinitions, args.argv);
+      this.validateParameters(command, this.commandDefinitions, args.argv);
 
       switch (command) {
         case constants.CMD_ADD:
@@ -114,7 +114,7 @@ class Registry extends BaseCLIHandler {
           return this.help([constants.CMD_LIST])
       }
     } catch (error) {
-      AppHelper.handleCLIError(error);
+      this.handleCLIError(error);
     }
   }
